refactor(materials): move files
