+++ packages/genui-renderer/src/core/index.ts
@@ -1,6 +1,6 @@
-export * from './materials';
+export * from '../../../core/src/protocols/materials';
 export * from './schema';
 export * from './gen-prompt-config';
 export * from './component-types';
 export * from '../utils/extract-schema-json';
-export * from './chat';
+export * from '../../../core/src/protocols/chat';
