+++ packages/materials/angular-opentiny-ng/example-schema.ts
