+++ packages/materials/angular-opentiny-ng/package.json
@@ -1,4 +1,7 @@
 {
     "name": "@opentiny/genui-sdk-materials-angular-opentiny-ng",
-    "version": "0.0.1"
+    "version": "0.0.1",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "type": "module"
 }
