+++ packages/materials/vue-opentiny-vue/chart.json
