+++ packages/materials/vue-opentiny-vue/example-schema.ts
